Move `ImageCommentsUIComposer` and `MainQueueDispatchDecorator` to production
