+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -12,33 +12,6 @@ import UIKit
 import EssentialFeed
 import EssentialFeediOS
 
-
-final class ImageCommentsUIComposer {
-	static func imageCommentsComposedWith(imageCommentsLoader: ImageCommentsLoader) -> ImageCommentsViewController {
-		let presentationAdapter = ImageCommentsPresentationAdapter(imageLoader: MainQueueDispatchDecorator(decoratee: imageCommentsLoader))
-		
-		let commentsController = makeImageCommentsViewController(
-			delegate: presentationAdapter,
-			title: ImageCommentsPresenter.title)
-		
-		presentationAdapter.presenter = ImageCommentsPresenter(
-			commentsView: ImageCommentsViewAdapter(controller: commentsController),
-			loadingView: WeakRefVirtualProxy(commentsController),
-			errorView: WeakRefVirtualProxy(commentsController))
-		
-		return commentsController
-	}
-	
-	private static func makeImageCommentsViewController(delegate: ImageCommentsViewControllerDelegate, title: String) -> ImageCommentsViewController {
-		let bundle = Bundle(for: ImageCommentsViewController.self)
-		let storyboard = UIStoryboard(name: "ImageComments", bundle: bundle)
-		let imageController = storyboard.instantiateInitialViewController() as! ImageCommentsViewController
-		imageController.delegate = delegate
-		imageController.title = title
-		return imageController
-	}
-}
-
 final class ImageCommentsUIIntegrationTests: XCTestCase {
 	
 	func test_imageCommentsView_hasTitle() {
@@ -289,25 +262,3 @@ private extension ImageCommentsViewController {
 	}
 }
 
-final class MainQueueDispatchDecorator<T> {
-	private let decoratee: T
-	
-	init(decoratee: T) {
-		self.decoratee = decoratee
-	}
-	
-	func dispatch(completion: @escaping () -> Void) {
-		guard Thread.isMainThread else {
-			return DispatchQueue.main.async(execute: completion)
-		}
-		completion()
-	}
-}
-
-extension MainQueueDispatchDecorator: ImageCommentsLoader where T == ImageCommentsLoader {
-	func loadImageComments(completion: @escaping (ImageCommentsLoader.Result) -> Void) -> ImageCommmentsLoaderTask {
-		decoratee.loadImageComments { [weak self] result in
-			self?.dispatch { completion(result) }
-		}
-	}
-}
